Worker group expansion panel
