+++ frontend/src/components/ShootWorkers/WorkerGroup.vue
@@ -75,10 +75,9 @@ import VendorIcon from '@/components/VendorIcon'
 import CodeBlock from '@/components/CodeBlock'
 import find from 'lodash/find'
 import get from 'lodash/get'
-import map from 'lodash/map'
-import join from 'lodash/join'
 import chunk from 'lodash/chunk'
 import { mapGetters } from 'vuex'
+import { workerGroupData } from '@/utils'
 
 export default {
   name: 'worker-group',
@@ -123,192 +122,7 @@ export default {
       return find(machineImages, { name, version })
     },
     workerGroupData () {
-      const data = []
-
-      const machineItem = {
-        title: 'Machine',
-        icon: 'mdi-server',
-        items: [
-          {
-            title: 'Type',
-            value: get(this.workerGroup, 'machine.type')
-          }
-        ]
-      }
-      const machineArchitecture = this.workerGroup.machine.architecture
-      if (machineArchitecture) {
-        machineItem.items.push(
-          {
-            title: 'Architecture',
-            value: machineArchitecture
-          }
-        )
-      }
-      const machineType = this.machineType
-      if (machineType) {
-        machineItem.items.push(...[
-          {
-            title: 'CPUs',
-            value: machineType.cpu
-          },
-          {
-            title: 'GPUs',
-            value: machineType.gpu
-          },
-          {
-            title: 'Memory',
-            value: machineType.memory
-          }
-        ])
-      }
-      const zones = this.workerGroup.zones
-      if (zones?.length) {
-        machineItem.items.push(
-          {
-            title: 'Zones',
-            value: join(zones, ', ')
-          }
-        )
-      }
-      data.push(machineItem)
-
-      const { name, version } = get(this.workerGroup, 'machine.image', {})
-      const imageItem = {
-        title: 'Image',
-        icon: 'mdi-disc',
-        items: [
-          {
-            title: 'Name',
-            value: name
-          },
-          {
-            title: 'Version',
-            value: version
-          }
-        ]
-      }
-      const machineImage = this.machineImage
-      if (!machineImage) {
-        imageItem.items.push({
-          title: 'Attention',
-          value: 'Image not found in cloud profile'
-        })
-      } else {
-        if (machineImage.expirationDate) {
-          imageItem.items.push({
-            title: 'Expires',
-            value: machineImage.expirationDateString
-          })
-        }
-      }
-      data.push(imageItem)
-
-      const cri = this.workerGroup.cri
-      if (cri) {
-        const criItem = {
-          title: 'Container Runtime',
-          icon: 'mdi-oci',
-          items: [
-            {
-              title: 'Name',
-              value: cri.name
-            }
-          ]
-        }
-        if (cri.containerRuntimes?.length) {
-          const containerRuntimes = map(cri.containerRuntimes, 'type')
-          criItem.items.push({
-            title: 'Additional OCI Runtimes',
-            value: join(containerRuntimes, ', ')
-          })
-        }
-        data.push(criItem)
-      }
-
-      const storage = get(this.machineType, 'storage', {})
-      const volume = get(this.workerGroup, 'volume', {})
-
-      // all infrastructures support volume sizes, but for some they are optional
-      // if no size is defined on the worker itself, check if machine storage defines a default size
-      const volumeSize = volume.size || storage.size
-
-      // workers with volume type (e.g. aws)
-      if (volume.type) {
-        const volumeItem = {
-          title: 'Volume',
-          icon: 'mdi-harddisk',
-          items: [
-            {
-              title: 'Type',
-              value: volume.type
-            }
-          ]
-        }
-        const volumeType = this.volumeType
-        if (volumeType) {
-          volumeItem.items.push({
-            title: 'Class',
-            value: volumeType.class
-          })
-        }
-        if (volumeSize) {
-          volumeItem.items.push({
-            title: 'Size',
-            value: volumeSize
-          })
-        }
-        data.push(volumeItem)
-      }
-
-      // workers with storage in machine type metadata (e.g. openstack)
-      if (storage.type) {
-        const storageItem = {
-          title: 'Storage',
-          icon: 'mdi-harddisk',
-          items: [
-            {
-              title: 'Type',
-              value: storage.type
-            },
-            {
-              title: 'Class',
-              value: storage.class
-            }
-          ]
-        }
-        if (volumeSize) {
-          storageItem.items.push({
-            title: 'Size',
-            value: volumeSize
-          })
-        }
-        data.push(storageItem)
-      }
-
-      const autoscalerItem = {
-        title: 'Autoscaler',
-        icon: 'mdi-chart-line-variant',
-        items: [
-          {
-            title: 'Maximum',
-            value: this.workerGroup.maximum
-          },
-          {
-            title: 'Minimum',
-            value: this.workerGroup.minimum
-          },
-          {
-            title: 'Max. Surger',
-            value: this.workerGroup.maxSurge
-          },
-          {
-            title: 'Max. Unavailable',
-            value: this.workerGroup.maxUnavailable
-          }
-        ]
-      }
-
-      data.push(autoscalerItem)
+      const data = workerGroupData(this.machineType, this.machineImage, this.volumeType, this.workerGroup)
       return chunk(data, Math.ceil(data.length / 2))
     },
     machineImageIcon () {
